Moved parsers to their own namespace
